+++ libs/base_core/src/code/mod.rs
@@ -6,6 +6,7 @@
 #![cfg_attr(base_safe_code, forbid(unsafe_code))]
 
 mod any; // dynamic typing and reflection
+mod guard; // ScopeGuard
 mod reexports;
 
 pub mod error; // general errors definitions
@@ -17,7 +18,7 @@ pub mod util; // utility macros and functions
 
 util::structural_mods! { // _mods, _pub_mods, _workspace_internals
     _mods {
-        pub use super::{any::*, reexports::*};
+        pub use super::{any::*, guard::*, reexports::*};
     }
     _pub_mods {
         pub use super::{