+++ docs/changelog.md
@@ -123,6 +123,8 @@ Many feature gates are removed in order to make most features make always availa
 -----------
 
 ## code
+- move `ScopeGuard` to [base_core].
+
 ### error
 - update `define_error!` macro.
   - move to `code::error`.